refactor: 코드 정렬 및 사용하지 않는 클래스 삭제
