+++ src/main/java/domain/BlackjackResult.java
@@ -4,7 +4,6 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class BlackjackResult {
-
     private final Map<Player, BetResult> playerBets = new LinkedHashMap<>();
 
     private BlackjackResult() {
