+++ src/main/java/domain/Card.java
@@ -1,7 +1,6 @@
 package domain;
 
 public class Card {
-
     private final Rank rank;
     private final Suit suit;
