+++ src/main/java/domain/Cards.java
@@ -5,7 +5,6 @@ import java.util.Collections;
 import java.util.List;
 
 public class Cards {
-
     private final List<Card> cards;
 
     private Cards(List<Card> cards) {
