+++ src/main/java/domain/GameResult.java
@@ -1,7 +1,6 @@
 package domain;
 
 public enum GameResult {
-
     WIN("승"),
     LOSE("패"),
     DRAW("무"),
