--- src/main/java/domain/NoShuffleStrategy.java
@@ -1,7 +1,0 @@
-package domain;
-
-public class NoShuffleStrategy implements CardShuffleStrategy{
-    @Override
-    public void shuffle(Cards cards) {
-    }
-}
