+++ src/main/java/domain/Rank.java
@@ -1,7 +1,6 @@
 package domain;
 
 public enum Rank {
-
     TWO(2, "2"),
     THREE(3, "3"),
     FOUR(4, "4"),
