+++ src/main/java/domain/Suit.java
@@ -1,7 +1,6 @@
 package domain;
 
 public enum Suit {
-
     SPADE("스페이드"),
     CLOVER("클로버"),
     DIAMOND("다이아몬드"),
