+++ src/main/java/domain/shuffle/CardShuffleStrategy.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.shuffle;
+
+import domain.Cards;
 
 public interface CardShuffleStrategy {
     void shuffle(Cards cards);
