+++ src/main/java/view/InputView.java
@@ -1,7 +1,7 @@
 package view;
 
 import domain.Player;
-import meesage.InputMessage;
+import view.mesage.InputMessage;
 
 import java.util.Scanner;
 
@@ -9,12 +9,12 @@ public class InputView {
 
     public static final Scanner scanner = new Scanner(System.in);
 
-    public String askUsersName(){
+    public String askUsersName() {
         System.out.println(InputMessage.ASK_USER_NAME.getMessage());
         return scanner.nextLine();
     }
 
-    public String askAddCard(String name){
+    public String askAddCard(String name) {
         System.out.printf(InputMessage.ASK_ADD_CARD.getMessage(), name);
         System.out.println();
         return scanner.nextLine();
