+++ src/main/java/view/mesage/InputMessage.java
@@ -1,7 +1,6 @@
-package meesage;
+package view.mesage;
 
 public enum InputMessage {
-
     ASK_USER_NAME("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)"),
     ASK_ADD_CARD("%s는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)"),
     USER_INPUT_YES("y"),
