+++ src/test/java/domain/shuffle/NoShuffleStrategy.java
@@ -1,0 +1,9 @@
+package domain.shuffle;
+
+import domain.Cards;
+
+public class NoShuffleStrategy implements CardShuffleStrategy {
+    @Override
+    public void shuffle(Cards cards) {
+    }
+}
